third_party/blink/web_tests/editing/execCommand/delete-empty-container-expected.txt

CONSOLE WARNING: document.execCommand() doesn't work with an invalid HTML structure. It is corrected automatically.
Deleting an empty selection should have no effect.
